+++ src/pages/user/Submit.js
@@ -1,25 +1,22 @@
 import React from 'react';
-import Logo from '../components/Logo';
-import Navigation from '../components/Navigation';
-import Formulaire from '../components/Formulaire';
+import Logo from '../../components/Logo';
+import NavigationUserComponent from '../../components/navigation/NavigationUser.component';
+import FormulaireComponent from '../../components/user/Formulaire.component';
 
 
 const Submit = () => {
     return (
         <div>
-            <Navigation />
+            <NavigationUserComponent />
             <Logo />
 
             <br />
             <br />
             <br />
-            
             <div className='divCorpSubmit'>
-
-
                 <h1 className='titleSubmit'>Submit new coin to Fantom Tribune</h1>
                 <br />
-                <Formulaire />
+                <FormulaireComponent/>
 
             </div>
